Pull #no_fanout into reusable concern
